app/requirements.txt
SQLAlchemy==2.0.23
alembic==1.13.1
pymysql==1.1.0
python-dotenv==1.0.0